Initial Commit for JMM Site
This creates the majority of pages (excluding the links page) for the
page. Styling is finished for desktop clients, but not yet responsive
as the design changes. Content will be modified at the same time.

A robots.txt file has been added to the repo, to be replaced with a
CNAME file once the site is no longer being staged on github pages'
domain and is live.